# backend/server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: easy-admin Server API
  version: 0.4.0
  description: |
    easy-admin Go service API. Health probes plus administrator authentication,
    RBAC management, read-only audit APIs, and the business-user platform
    (public settings, registration, verification, authentication, and
    administrator-managed users/levels/points/system settings). All responses
    use the shared JSON envelope with a request ID. Persistent IDs are decimal
    strings; timestamps are UTC RFC 3339; point values are fixed four-decimal
    strings (for example "123.4567") to avoid JavaScript precision loss.
servers:
  - url: http://127.0.0.1:8080
    description: Local development
...
paths:
...
  /api/v1/public/settings:
    get:
      tags: [public-settings]
      summary: Public platform settings whitelist
      operationId: getPublicSettings
      parameters:
        - $ref: '#/components/parameters/RequestID'
      responses:
        '200':
          description: Safe public settings (no SMTP secrets, no keys)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSettingsSuccess'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: |
        Optional inbound request ID. Accepted only when 1-128 characters and
        every character is ASCII alphanumeric or one of `.`, `_`, `:`, `-`.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]{1,128}$'
...
  responses:
...
    DependencyUnavailable:
      description: A required dependency is unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DependencyUnavailableError'
...
  schemas:
    Envelope:
      type: object
      required: [code, message, data, request_id]
      properties:
        code: { type: integer }
        message: { type: string }
        data: { type: object, additionalProperties: true }
        request_id: { type: string, minLength: 1 }
...
    PublicSettings:
      type: object
      required: [platform_name, public_frontend_url, public_api_url, registration_enabled, username_login_enabled, email_login_enabled, email_verification_required, default_avatar_url]
      properties:
        platform_name: { type: string }
        public_frontend_url: { type: string, format: uri }
        public_api_url: { type: string, format: uri }
        registration_enabled: { type: boolean }
        username_login_enabled: { type: boolean }
        email_login_enabled: { type: boolean }
        email_verification_required: { type: boolean }
        default_avatar_url: { type: string }
      additionalProperties: false
    PublicSettingsSuccess:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            code: { enum: [0] }
            data:
              $ref: '#/components/schemas/PublicSettings'
...
    DependencyUnavailableError:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            code: { type: integer, enum: [50001, 50003] }
            message: { type: string, enum: [dependency unavailable, smtp unavailable] }
            data: { type: object, additionalProperties: false }
...